+++ pennylane/capture/primitives.py
@@ -24,7 +24,7 @@ from pennylane._grad.value_and_grad import _get_value_and_grad_prim
 from pennylane._grad.vjp import _get_vjp_prim
 from pennylane.control_flow.for_loop import _get_for_loop_qfunc_prim
 from pennylane.control_flow.while_loop import _get_while_loop_qfunc_prim
-from pennylane.core.operator import _get_abstract_operator
+from pennylane.core.operator.base import _get_abstract_operator  # tach-ignore
 from pennylane.measurements.capture_measurements import _get_abstract_measurement
 from pennylane.ops.mid_measure.mid_measure import _create_mid_measure_primitive
 from pennylane.ops.mid_measure.pauli_measure import _create_pauli_measure_primitive
