+++ pennylane/core/operator/cv.py
@@ -1,0 +1,299 @@
+# Copyright 2018-2026 Xanadu Quantum Technologies Inc.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+Base class for CV Operators.
+"""
+
+import numpy as np
+
+from pennylane import math
+from pennylane._class_property import classproperty
+from pennylane.queuing import QueuingManager
+
+from .base import Operation, Operator
+
+
+class CV:
+    """A mixin base class denoting a continuous-variable operation."""
+
+    # pylint: disable=no-member
+
+    def heisenberg_expand(self, U, wire_order):
+        """Expand the given local Heisenberg-picture array into a full-system one.
+
+        Args:
+            U (array[float]): array to expand (expected to be of the dimension ``1+2*self.num_wires``)
+            wire_order (Wires): global wire order defining which subspace the operator acts on
+
+        Raises:
+            ValueError: if the size of the input matrix is invalid or `num_wires` is incorrect
+
+        Returns:
+            array[float]: expanded array, dimension ``1+2*num_wires``
+        """
+
+        U_dim = len(U)
+        nw = len(self.wires)
+
+        if U.ndim > 2:
+            raise ValueError("Only order-1 and order-2 arrays supported.")
+
+        if U_dim != 1 + 2 * nw:
+            raise ValueError(f"{self.name}: Heisenberg matrix is the wrong size {U_dim}.")
+
+        if len(wire_order) == 0 or len(self.wires) == len(wire_order):
+            # no expansion necessary (U is a full-system matrix in the correct order)
+            return U
+
+        if not wire_order.contains_wires(self.wires):
+            raise ValueError(
+                f"{self.name}: Some observable wires {self.wires} do not exist on this device with wires {wire_order}"
+            )
+
+        # get the indices that the operation's wires have on the device
+        wire_indices = wire_order.indices(self.wires)
+
+        # expand U into the I, x_0, p_0, x_1, p_1, ... basis
+        dim = 1 + len(wire_order) * 2
+
+        def loc(w):
+            "Returns the slice denoting the location of (x_w, p_w) in the basis."
+            ind = 2 * w + 1
+            return slice(ind, ind + 2)
+
+        if U.ndim == 1:
+            W = np.zeros(dim)
+            W[0] = U[0]
+            for k, w in enumerate(wire_indices):
+                W[loc(w)] = U[loc(k)]
+        elif U.ndim == 2:
+            W = np.zeros((dim, dim)) if isinstance(self, CVObservable) else np.eye(dim)
+            W[0, 0] = U[0, 0]
+
+            for k1, w1 in enumerate(wire_indices):
+                s1 = loc(k1)
+                d1 = loc(w1)
+
+                # first column
+                W[d1, 0] = U[s1, 0]
+                # first row (for gates, the first row is always (1, 0, 0, ...), but not for observables!)
+                W[0, d1] = U[0, s1]
+
+                for k2, w2 in enumerate(wire_indices):
+                    W[d1, loc(w2)] = U[s1, loc(k2)]  # block k1, k2 in U goes to w1, w2 in W.
+        return W
+
+    @staticmethod
+    def _heisenberg_rep(p):
+        r"""Heisenberg picture representation of the operation.
+
+        * For Gaussian CV gates, this method returns the matrix of the linear
+          transformation carried out by the gate for the given parameter values.
+          The method is not defined for non-Gaussian gates.
+
+          **The existence of this method is equivalent to setting** ``grad_method = 'A'``.
+
+        * For observables, returns a real vector (first-order observables) or
+          symmetric matrix (second-order observables) of expansion coefficients
+          of the observable.
+
+        For single-mode Operations we use the basis :math:`\mathbf{r} = (\I, \x, \p)`.
+        For multi-mode Operations we use the basis :math:`\mathbf{r} = (\I, \x_0, \p_0, \x_1, \p_1, \ldots)`.
+
+        .. note::
+
+            For gates, we assume that the inverse transformation is obtained
+            by negating the first parameter.
+
+        Args:
+            p (Sequence[float]): parameter values for the transformation
+
+        Returns:
+            array[float]: :math:`\tilde{U}` or :math:`q`
+        """
+        # pylint: disable=unused-argument
+        return None
+
+    @classproperty
+    def supports_heisenberg(self):
+        """Whether a CV operator defines a Heisenberg representation.
+
+        This indicates that it is Gaussian and does not block the use
+        of the parameter-shift differentiation method if found between the differentiated gate
+        and an observable.
+
+        Returns:
+            boolean
+        """
+        return CV._heisenberg_rep != self._heisenberg_rep
+
+
+class CVOperation(CV, Operation):
+    """Base class representing continuous-variable quantum gates.
+
+    CV operations provide a special Heisenberg representation, as well as custom methods
+    for differentiation.
+
+    Args:
+        params (tuple[tensor_like]): trainable parameters
+        wires (Iterable[Any] or Any): Wire label(s) that the operator acts on.
+            If not given, args[-1] is interpreted as wires.
+    """
+
+    @classproperty
+    def supports_parameter_shift(self):
+        """Returns True iff the CV Operation supports the parameter-shift differentiation method.
+        This means that it has ``grad_method='A'`` and
+        has overridden the :meth:`~.CV._heisenberg_rep` static method.
+        """
+        return self.grad_method == "A" and self.supports_heisenberg
+
+    def heisenberg_pd(self, idx):
+        """Partial derivative of the Heisenberg picture transform matrix.
+
+        Computed using grad_recipe.
+
+        Args:
+            idx (int): index of the parameter with respect to which the
+                partial derivative is computed.
+        Returns:
+            array[float]: partial derivative
+        """
+        # get the gradient recipe for this parameter
+        recipe = self.grad_recipe[idx]
+
+        # Default values
+        multiplier = 0.5
+        a = 1
+        shift = np.pi / 2
+
+        # We set the default recipe to as follows:
+        # ∂f(x) = c*f(x+s) - c*f(x-s)
+        default_param_shift = [[multiplier, a, shift], [-multiplier, a, -shift]]
+        param_shift = default_param_shift if recipe is None else recipe
+
+        pd = None  # partial derivative of the transformation
+
+        p = self.parameters
+
+        original_p_idx = p[idx]
+        for c, _a, s in param_shift:
+            # evaluate the transform at the shifted parameter values
+            p[idx] = _a * original_p_idx + s
+            U = self._heisenberg_rep(p)
+
+            if pd is None:
+                pd = c * U
+            else:
+                pd += c * U
+
+        return pd
+
+    def heisenberg_tr(self, wire_order, inverse=False):
+        r"""Heisenberg picture representation of the linear transformation carried
+        out by the gate at current parameter values.
+
+        Given a unitary quantum gate :math:`U`, we may consider its linear
+        transformation in the Heisenberg picture, :math:`U^\dagger(\cdot) U`.
+
+        If the gate is Gaussian, this linear transformation preserves the polynomial order
+        of any observables that are polynomials in :math:`\mathbf{r} = (\I, \x_0, \p_0, \x_1, \p_1, \ldots)`.
+        This also means it maps :math:`\text{span}(\mathbf{r})` into itself:
+
+        .. math:: U^\dagger \mathbf{r}_i U = \sum_j \tilde{U}_{ij} \mathbf{r}_j
+
+        For Gaussian CV gates, this method returns the transformation matrix for
+        the current parameter values of the Operation. The method is not defined
+        for non-Gaussian (and non-CV) gates.
+
+        Args:
+            wire_order (Wires): global wire order defining which subspace the operator acts on
+            inverse  (bool): if True, return the inverse transformation instead
+
+        Raises:
+            RuntimeError: if the specified operation is not Gaussian or is missing the `_heisenberg_rep` method
+
+        Returns:
+            array[float]: :math:`\tilde{U}`, the Heisenberg picture representation of the linear transformation
+        """
+        p = [math.toarray(a) for a in self.parameters]
+        if inverse:
+            try:
+                # TODO: expand this for the new par domain class, for non-unitary matrices.
+                p[0] = np.linalg.inv(p[0])
+            except np.linalg.LinAlgError:
+                p[0] = -p[0]  # negate first parameter
+        U = self._heisenberg_rep(p)
+
+        # not defined?
+        if U is None:
+            raise RuntimeError(
+                f"{self.name} is not a Gaussian operation, or is missing the _heisenberg_rep method."
+            )
+
+        return self.heisenberg_expand(U, wire_order)
+
+
+class CVObservable(CV, Operator):
+    r"""Base class representing continuous-variable observables.
+
+    CV observables provide a special Heisenberg representation.
+
+    The class attribute :attr:`~.ev_order` can be defined to indicate
+    to PennyLane whether the corresponding CV observable is a polynomial in the
+    quadrature operators. If so,
+
+    * ``ev_order = 1`` indicates a first order polynomial in quadrature
+      operators :math:`(\x, \p)`.
+
+    * ``ev_order = 2`` indicates a second order polynomial in quadrature
+      operators :math:`(\x, \p)`.
+
+    If :attr:`~.ev_order` is not ``None``, then the Heisenberg representation
+    of the observable should be defined in the static method :meth:`~.CV._heisenberg_rep`,
+    returning an array of the correct dimension.
+
+    Args:
+       params (tuple[tensor_like]): trainable parameters
+       wires (Iterable[Any] or Any): Wire label(s) that the operator acts on.
+           If not given, args[-1] is interpreted as wires.
+    """
+
+    is_verified_hermitian = True
+
+    def queue(self, context=QueuingManager):
+        """Avoids queuing the observable."""
+        return self
+
+    ev_order = None  #: None, int: Order in `(x, p)` that a CV observable is a polynomial of.
+
+    def heisenberg_obs(self, wire_order):
+        r"""Representation of the observable in the position/momentum operator basis.
+
+        Returns the expansion :math:`q` of the observable, :math:`Q`, in the
+        basis :math:`\mathbf{r} = (\I, \x_0, \p_0, \x_1, \p_1, \ldots)`.
+
+        * For first-order observables returns a real vector such
+          that :math:`Q = \sum_i q_i \mathbf{r}_i`.
+
+        * For second-order observables returns a real symmetric matrix
+          such that :math:`Q = \sum_{ij} q_{ij} \mathbf{r}_i \mathbf{r}_j`.
+
+        Args:
+            wire_order (Wires): global wire order defining which subspace the operator acts on
+        Returns:
+            array[float]: :math:`q`
+        """
+        p = self.parameters
+        U = self._heisenberg_rep(p)
+        return self.heisenberg_expand(U, wire_order)
