+++ pennylane/operation.py
@@ -16,8 +16,10 @@ The former location of pennylane/core/operator.
 
 import pennylane as qp
 from pennylane import math
-from pennylane.core.operator import *  # pylint: disable=wildcard-import, unused-wildcard-import
-from pennylane.core.operator import _UNSET_BATCH_SIZE  # pylint: disable=unused-import
+from pennylane.core.operator import *  # pylint: disable=wildcard-import, unused-wildcard-import # tach-ignore
+from pennylane.core.operator.base import (  # pylint: disable=unused-import # tach-ignore
+    _UNSET_BATCH_SIZE,
+)
 from pennylane.exceptions import (  # pylint: disable=unused-import
     AdjointUndefinedError,
     DecompositionUndefinedError,
@@ -32,7 +34,7 @@ from pennylane.exceptions import (  # pylint: disable=unused-import
     TermsUndefinedError,
 )
 from pennylane.typing import TensorLike
-from pennylane.wires import Wires  # pylint: disable-unused-import
+from pennylane.wires import Wires  # pylint: disable=unused-import
 
 
 def operation_derivative(operation: Operation) -> TensorLike:
