+++ tach.toml
@@ -337,6 +337,10 @@ cannot_depend_on = ["pennylane.ftqc", "pennylane.labs"]
 path = "pennylane.core.operator"
 cannot_depend_on = ["pennylane.ftqc", "pennylane.labs"]
 
+[[interfaces]]
+expose = ["Operator", "Operation", "CV", "CVObservable", "CVOperation", "Channel", "StatePrepBase"]
+from = ["pennylane.core.operator"]
+
 [[modules]]
 path = "pennylane.ops"
 cannot_depend_on = ["pennylane.ftqc", "pennylane.labs"]
